requirements.txt
Office365-REST-Python-Client>=2.5.0
pywin32>=306
python-dotenv>=1.0.0
